Removing warnings, mostly signed vs unsigned. Since code uses ArrayViews, need to declare their array indices as ArrayView<template_type>::size_type.

+++ packages/zoltan2/util/xmlToHeaderDefinition.cpp
@@ -72,7 +72,7 @@ static string fixLine(string oldLine)
   string newLine;
   string quotedQuote("\\\"");
 
-  for (int i=0; i < oldLine.size(); i++){
+  for (size_t i=0; i < oldLine.size(); i++){
     char c = oldLine[i];
     if (c == quote[0])
       newLine.append(quotedQuote);
